+++ test/benchmark/ArraysPerformanceTest.ts
@@ -6,6 +6,8 @@ declare var process: { hrtime(time?: [number, number]): [number, number] };
 var asr = chai.assert;
 
 suite("ArraysPerformance", function ArraysPerformance() {
+    this.timeout(3000);
+
     var res = 0;
     // warmup
     res += copyLoop(200, [1, 2, 3]);